✨ feat(batch): 벡터 업로드 배치(디스패처·종료·러너) 및 일일 AI 배치 스케줄러 추가

+++ app-api/src/main/java/com/tasteam/infra/ai/dto/AiVectorUploadRequest.java
@@ -8,14 +8,14 @@ public record AiVectorUploadRequest(
 	List<ReviewPayload> reviews,
 	List<RestaurantPayload> restaurants) {
 	public record ReviewPayload(
-		Integer id,
+		long id,
 		@JsonProperty("restaurant_id")
 		long restaurantId,
 		String content) {
 	}
 
 	public record RestaurantPayload(
-		Integer id,
+		long id,
 		String name) {
 	}
 }
